LayoutTests/http/tests/appcache/abort-cache-ondownloading-resource-404-expected.txt

CONSOLE MESSAGE: Application Cache download process was aborted.
This tests that download process was aborted after downloading event although resource was not found.
SUCCESS
